Separating tests and integration_tests modules
