+++ canary.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e # everything must pass
 . .upgrade-deps.sh # upgrade all deps to latest version
-. .unittest.sh # run the tests
+. .run-tests.sh # run the tests
